+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
     "crates/perl-lexer",
     "crates/perl-parser",
     "crates/perl-parser-core",
+    "crates/perl-parser-recovery",
     "crates/perl-parser-pest",
     "crates/perl-semantic-analyzer",
     "crates/perl-workspace-index",
@@ -157,6 +158,7 @@ allow = [
 
     # Tier 2 — Core infrastructure
     "perl-parser-core",
+    "perl-parser-recovery",
     "perl-tdd-support",
     "perl-edit",
     "perl-source-editing",
@@ -352,6 +354,7 @@ perl-lsp-capability-map = { path = "crates/perl-lsp-capability-map", version = "
 
 # Tier 2: Single-level dependencies
 perl-parser-core = { path = "crates/perl-parser-core", version = "0.11.0" }
+perl-parser-recovery = { path = "crates/perl-parser-recovery", version = "0.11.0" }
 perl-lsp-transport = { path = "crates/perl-lsp-transport", version = "0.11.0" }
 perl-lsp-tooling = { path = "crates/perl-lsp-tooling", version = "0.11.0" }
 perl-lsp-on-type-formatting = { path = "crates/perl-lsp-on-type-formatting", version = "0.11.0" }
