+++ crates/perl-parser-core/Cargo.toml
@@ -34,6 +34,7 @@ perl-regex = { workspace = true }
 perl-heredoc = { workspace = true }
 perl-error = { workspace = true }
 perl-tokenizer = { workspace = true }
+perl-parser-recovery = { workspace = true }
 
 [dev-dependencies]
 perl-tdd-support = { workspace = true }
